requirements.txt
flask
flask-socketio
pyqt6
pyqt6-webengine
canopen
python-can
alpinejs
